moving config.json to config folder
